+++ src/threedidepth/commands.py
@@ -12,8 +12,8 @@ from threedidepth.calculate import MODE_LIZARD
 from threedidepth.calculate import MODE_LIZARD_VAR
 
 # keep old constants for backwards compatibility
-from threedidepth.calculate import MODE_CONSTANT_S1
-from threedidepth.calculate import MODE_LIZARD_S1
+from threedidepth.calculate import MODE_CONSTANT_S1  # noqa
+from threedidepth.calculate import MODE_LIZARD_S1  # noqa
 
 
 def threedidepth(*args):